Print sizes on this machine

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int c;
 	float f;
 
-	printf("Size of char: %lu byte(s)\n", sizeof(a));
-	printf("Size of int: %lu byte(s)\n", sizeof(i));
-	printf("Size of long int: %lu byte(s)\n", sizeof(b));
-	printf("Size of long long int: %lu byte(s)\n", sizeof(c));
-	printf("Size of float: %lu byte(s)\n", sizeof(f));
+	printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(a));
+	printf("Size of int: %lu byte(s)\n", (unsigned long)sizeof(i));
+	printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(f));
 return (0);
